Added new fixes

+++ testfunc.hpp
@@ -1,11 +1,13 @@
 #ifndef _TESTFUNC_H_
 #define _TESTFUNC_H_
 
+#include <cstddef>
+
 class TestFunc {
 public:
-    int dim;
+    size_t dim;
 
-    TestFunc(int idim) {
+    TestFunc(size_t idim) {
         dim = idim;
     };
 
@@ -16,9 +18,9 @@ class TestFuncBounds: public TestFunc {
 public:
     double *x_bound_min;
     double *x_bound_max;
-    int max_num_evaluations;
+    size_t max_num_evaluations;
 
-    TestFuncBounds(int dim, int g_max_num_evaluations): TestFunc(dim) {
+    TestFuncBounds(size_t dim, size_t g_max_num_evaluations): TestFunc(dim) {
         x_bound_min = new double[dim];
         x_bound_max = new double[dim];
         max_num_evaluations = g_max_num_evaluations;
